% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_coefficient.R
\name{boundary_coefficient}
\alias{boundary_coefficient}
\title{Compute the boundary coefficients for a weighted graph}
\usage{
boundary_coefficient(G, pairD = NULL)
}
\arguments{
\item{G}{the graph for which to compute the boundary measure of each node}

\item{pairD}{pairwise distance object of G, computed if missing}
}
\value{
a vector of length |V(G)| containing the boundary coefficients
}
\description{
Compute the boundary coefficients for a weighted graph
}
