% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_hop_pairG.R
\name{bounded_hop_pairG}
\alias{bounded_hop_pairG}
\title{Compute the weighted distances between nodes if there unweighted distance is less than a given integer k
This code is more memory efficient than computing the full paiwise distance matrix and discarding unnecessary nodes,
but may run slower due to the optimized implementation of igraphs distances function in C}
\usage{
bounded_hop_pairG(G, k = 2)
}
\arguments{
\item{G}{the graph for which to compute the}

\item{k}{maximum allowd hops between pairs of nodes for which the weighted distance is to be computed}
}
\value{
a sparse matrix storing the weighted distance D_uv between nodes u and v if the unweighted distance
        between u and v is less than k
}
\description{
Compute the weighted distances between nodes if there unweighted distance is less than a given integer k
This code is more memory efficient than computing the full paiwise distance matrix and discarding unnecessary nodes,
but may run slower due to the optimized implementation of igraphs distances function in C
}
