% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elbow.R
\name{get_elbow}
\alias{get_elbow}
\title{Compute the elbow(s) for curve(s), the function is specifically designed to process the curves outputed from the search pruning}
\usage{
get_elbow(C, alpha = NA)
}
\arguments{
\item{C}{data frame or matrix containing the x-coordinates (col1) and y-coordinates (col2) of the curve(s)
optionally, a third column may be provided to group the coordinates according to different curves (col3)}

\item{alpha}{optional threshold between 0 and 1, denoting a sensitivity to adding new leaves
if not provided, the elbow is computed by minimizing the absolute central differences}
}
\value{
integer vector containing the computed elbow index for each curve
}
\description{
Compute the elbow(s) for curve(s), the function is specifically designed to process the curves outputed from the search pruning
}
