% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leave_combs.R
\name{get_leave_combs}
\alias{get_leave_combs}
\title{Compute possible leave configurations across multiple components for a given number of leaves}
\usage{
get_leave_combs(c, l)
}
\arguments{
\item{c}{number of components}

\item{l}{number of leaves}
}
\value{
a matrix off c columns where each row is a possible combination of leaves across the c components
}
\description{
Compute possible leave configurations across multiple components for a given number of leaves
}
