% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_leaves.R
\name{get_new_leaves}
\alias{get_new_leaves}
\title{Get a new backbone with a different amount of leaves
This method only works when the backbone is returned using the 'search' method to prune}
\usage{
get_new_leaves(backbone, leaves, G = NULL, assign = FALSE,
  prob = TRUE, col = TRUE)
}
\arguments{
\item{backbone}{a backbone object to be converted to a new backbone object with a different amoubt of leaves}

\item{leaves}{vector of integers, denoting the number of leaves to be included for each component treated separately in the pine}

\item{G}{the original graph on top of the backbone, standard set to NULL}

\item{assign}{whether to produce a branch assignment of the resulting new backbone, standard set to FALSE}

\item{prob}{whether to assign nodes in G to paths according to a probability measure
if FALSE, then nodes are assigned to exactly one path, standard is TRUE}

\item{col}{whether to provide a coloring scheme for the paths and nodes, standard set to TRUE}
}
\value{
a backbone object with the new number of leaves
}
\description{
Get a new backbone with a different amount of leaves
This method only works when the backbone is returned using the 'search' method to prune
}
