% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_prune.R
\name{soft_prune}
\alias{soft_prune}
\title{Prune leaves in a graph either a given amount of times or until there are no more than a specified number of leaves
At each iteration, the leaves representing the fewest amount of nodes are pruned
Each node is set to represent only itself at the start of the iterations
When leaves are pruned, they are consequently represented by their sole neighbor}
\usage{
soft_prune(G, times = 1, leaves = NA)
}
\arguments{
\item{G}{graph to be pruned}

\item{times}{the number of times all leaves must be pruned, standard set to 1}

\item{leaves}{(maximum) number of leaves to be kept, 'times' is ignored if provided}
}
\value{
the pruned graph
}
\description{
Prune leaves in a graph either a given amount of times or until there are no more than a specified number of leaves
At each iteration, the leaves representing the fewest amount of nodes are pruned
Each node is set to represent only itself at the start of the iterations
When leaves are pruned, they are consequently represented by their sole neighbor
}
