% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_weights.R
\name{sum_weights}
\alias{sum_weights}
\title{Compute softmax weights for a graph using a predefined function}
\usage{
sum_weights(G, f)
}
\arguments{
\item{G}{Graph containing the edges for which the softmax values have to be computed}

\item{f}{vector of function values of length |V(G)|}
}
\value{
a vector containing the sum weight for each edge in G, ordered as E(G)
}
\description{
Compute softmax weights for a graph using a predefined function
}
