% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRgraph.R
\name{VRgraph}
\alias{VRgraph}
\title{Construct Vietoris-Rips graph from a point cloud data set by connecting all
nodes with a certain given distance eps}
\usage{
VRgraph(X, eps)
}
\arguments{
\item{X}{either a data.frame/matrix containing observations,
or a pairwise distance object to be used for constructing the complex}

\item{eps}{(strict) upper bound on edge lengths}
}
\value{
the Vietoris-Rips graph defined by X and distance parameter eps
}
\description{
Construct Vietoris-Rips graph from a point cloud data set by connecting all
nodes with a certain given distance eps
}
